+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## [0.1.1] - Border support
+## [0.1.1] - Small changes
+
+* Updated documentation
+* Added custom border radius
+* Able to change the animation duration
 * Added support for borders
 
 ## [0.1.0] - Background color fix